pyproject.toml
[project]
name = "moviemate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "crawl4ai>=0.8.5",
    "fastapi>=0.135.1",
    "langchain",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "seaborn>=0.13.2",
    "uvicorn>=0.42.0",
]
